+++ csds-client/client/clientV2.go
@@ -5,15 +5,16 @@ import (
 	"crypto/x509"
 	"errors"
 	"fmt"
+	"io"
+	"strings"
+	"time"
+
 	csdspb_v2 "github.com/envoyproxy/go-control-plane/envoy/service/status/v2"
 	envoy_type_matcher "github.com/envoyproxy/go-control-plane/envoy/type/matcher"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/oauth"
 	"google.golang.org/grpc/metadata"
-	"io"
-	"strings"
-	"time"
 )
 
 // ClientV2 implements the interface Client
@@ -149,6 +150,7 @@ func (c *ClientV2) Run() error {
 	}
 	defer c.clientConn.Close()
 
+	c.csdsClient = csdspb_v2.NewClientStatusDiscoveryServiceClient(c.clientConn)
 	var ctx context.Context
 	if c.metadata != nil {
 		ctx = metadata.NewOutgoingContext(context.Background(), c.metadata)
@@ -204,5 +206,84 @@ func (c *ClientV2) doRequest(streamClientStatus csdspb_v2.ClientStatusDiscoveryS
 		return err
 	}
 
+	return nil
+}
+
+// parseConfigStatus_v2 parses each xds config status to string
+func parseConfigStatus_v2(xdsConfig []*csdspb_v2.PerXdsConfig) []string {
+	var configStatus []string
+	for _, perXdsConfig := range xdsConfig {
+		status := perXdsConfig.GetStatus().String()
+		var xds string
+		if perXdsConfig.GetClusterConfig() != nil {
+			xds = "CDS"
+		} else if perXdsConfig.GetListenerConfig() != nil {
+			xds = "LDS"
+		} else if perXdsConfig.GetRouteConfig() != nil {
+			xds = "RDS"
+		} else if perXdsConfig.GetScopedRouteConfig() != nil {
+			xds = "SRDS"
+		}
+		if status != "" && xds != "" {
+			configStatus = append(configStatus, xds+"   "+status)
+		}
+	}
+	return configStatus
+}
+
+// printOutResponse_v2 processes response and print
+func printOutResponse_v2(response *csdspb_v2.ClientStatusResponse, opts ClientOptions) error {
+	if response.GetConfig() == nil || len(response.GetConfig()) == 0 {
+		fmt.Printf("No xDS clients connected.\n")
+		return nil
+	} else {
+		fmt.Printf("%-50s %-30s %-30s \n", "Client ID", "xDS stream type", "Config Status")
+	}
+
+	var hasXdsConfig bool
+
+	for _, config := range response.GetConfig() {
+		var id string
+		var xdsType string
+		if config.GetNode() != nil {
+			id = config.GetNode().GetId()
+			metadata := config.GetNode().GetMetadata().AsMap()
+
+			// control plane is expected to use "XDS_STREAM_TYPE" to communicate
+			// the stream type of the connected client in the response.
+			if metadata["XDS_STREAM_TYPE"] != nil {
+				xdsType = metadata["XDS_STREAM_TYPE"].(string)
+			}
+		}
+
+		if config.GetXdsConfig() == nil {
+			if config.GetNode() != nil {
+				fmt.Printf("%-50s %-30s %-30s \n", id, xdsType, "N/A")
+			}
+		} else {
+			hasXdsConfig = true
+
+			// parse config status
+			configStatus := parseConfigStatus_v2(config.GetXdsConfig())
+			fmt.Printf("%-50s %-30s ", id, xdsType)
+
+			for i := 0; i < len(configStatus); i++ {
+				if i == 0 {
+					fmt.Printf("%-30s \n", configStatus[i])
+				} else {
+					fmt.Printf("%-50s %-30s %-30s \n", "", "", configStatus[i])
+				}
+			}
+			if len(configStatus) == 0 {
+				fmt.Printf("\n")
+			}
+		}
+	}
+
+	if hasXdsConfig {
+		if err := printDetailedConfig(response, opts); err != nil {
+			return err
+		}
+	}
 	return nil
 }
